Move auxiliary classes into separate modules
