+++ scico/optimize/__init__.py
@@ -8,23 +8,9 @@
 """Optimization algorithms."""
 
 # isort: off
-from _admm import (
-    SubproblemSolver,
-    GenericSubproblemSolver,
-    LinearSubproblemSolver,
-    CircularConvolveSolver,
-    ADMM,
-)
+from _admm import ADMM
 from _ladmm import LinearizedADMM
-from _pgm import (
-    PGMStepSize,
-    BBStepSize,
-    AdaptiveBBStepSize,
-    LineSearchStepSize,
-    RobustLineSearchStepSize,
-    PGM,
-    AcceleratedPGM,
-)
+from _pgm import PGM, AcceleratedPGM
 from _primaldual import PDHG
 from _solver import minimize, minimize_scalar, cg
